Seperated KeyStoreEntry and made it a top level class.
